+++ astroquery/heasarc/tests/test_heasarc_remote.py
@@ -1,14 +1,16 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
-
+from time import time
 import pytest
 import requests
 
 from ...heasarc import Heasarc
 from ...utils import commons
 
+from .parametrization import parametrization_local_save_remote, patch_get, MockResponse
 
-@pytest.mark.remote_data
+
+@parametrization_local_save_remote
 class TestHeasarc:
     def test_custom_args(self):
         object_name = 'Crab'
@@ -76,7 +78,7 @@ class TestHeasarc:
         heasarc = Heasarc()
         response = heasarc.query_object_async(object_name, mission=mission)
         assert response is not None
-        assert type(response) is requests.models.Response
+        assert isinstance(response, (requests.models.Response, MockResponse))
 
     def test_query_region_async(self):
         heasarc = Heasarc()
@@ -85,7 +87,7 @@ class TestHeasarc:
         response = heasarc.query_region_async(c, mission=mission,
                                               radius='1 degree')
         assert response is not None
-        assert type(response) is requests.models.Response
+        assert isinstance(response, (requests.models.Response, MockResponse))
 
     def test_query_region(self):
         heasarc = Heasarc()